Create helper for positron data type
