+++ internal/services/api/post_packages_r_scan.go
@@ -8,7 +8,6 @@ import (
 	"io"
 	"net/http"
 	"path/filepath"
-	"strings"
 
 	"github.com/posit-dev/publisher/internal/inspect/dependencies/renv"
 	"github.com/posit-dev/publisher/internal/interpreters"
@@ -91,14 +90,3 @@ func (h *PostPackagesRScanHandler) ServeHTTP(w http.ResponseWriter, req *http.Re
 
 // repoOptsFromPositron converts inbound Positron settings to renv.RepoOptions.
 // Returns nil if no Positron R settings were provided.
-func repoOptsFromPositron(ps *positronSettings) *renv.RepoOptions {
-	if ps == nil || ps.R == nil {
-		return nil
-	}
-	mode := strings.TrimSpace(ps.R.DefaultRepositories)
-	ppm := strings.TrimSpace(ps.R.PackageManagerRepository)
-	return &renv.RepoOptions{
-		DefaultRepositories:      mode,
-		PackageManagerRepository: ppm,
-	}
-}
